+++ src/avatar/story.tsx
@@ -2,30 +2,31 @@ import React from 'react'
 
 import { storiesOf } from '@storybook/react'
 import { withInfo } from '@storybook/addon-info'
-import { withKnobs, text, boolean, select } from '@storybook/addon-knobs'
+import { withKnobs, boolean, text, select } from '@storybook/addon-knobs'
 
 import Avatar from 'avatar'
 
 const sizes = {
   default: '',
-  medium: 'medium',
-  large: 'large',
+  isMedium: 'medium',
+  isLarge: 'large',
 }
 
 const stories = storiesOf('Avatar', module)
 stories.addDecorator(withKnobs)
 
 stories.add(
-  'Overview',
+  'Basic',
   withInfo('')(() => {
     const modifiers = select('Size', sizes, 'default')
     const propsModifiers = { [modifiers]: !!modifiers }
     return (
       <Avatar
-        alt={text('alternative text', 'Avatar')}
-        checked={boolean('Id Check', false)}
         image={text('url', '//placehold.it/80x80')}
-        title={text('title for icon', 'id check')}
+        alt={text('alternative text', 'Avatar')}
+        isIdChecked={boolean('isIdChecked', false)}
+        unreadNotificationsCount={text('unreadNotificationsCount', '')}
+        unreadNotificationsCountAriaLabel={text('unreadNotificationsCountAriaLabel', '')}
         {...propsModifiers}
       />
     )