+++ src/icon/style.ts
@@ -1,0 +1,14 @@
+import css from 'styled-jsx/css'
+
+export default css`
+  .kirk-icon-wrapper {
+    display: inline-block;
+    position: relative;
+  }
+
+  :global(.kirk-icon-badge) {
+    position: absolute;
+    bottom: 12px;
+    left: 12px;
+  }
+`