+++ src/profile/index.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 import cc from 'classcat'
-import isEmpty from 'lodash.isempty'
 
-import Rating from 'rating'
-import Button from 'button'
 import Avatar from 'avatar'
+import Button from 'button'
+import Rating from 'rating'
 import ArrowIcon from 'icon/arrowIcon'
 
 import prefix from '_utils'
@@ -13,9 +12,9 @@ import style from 'profile/style'
 interface ProfileProps {
   readonly title: string,
   readonly className?: Classcat.Class,
-  readonly medium?: boolean,
+  readonly isMedium?: boolean,
   readonly picture?: string,
-  readonly checked?: boolean,
+  readonly isIdChecked?: boolean,
   readonly alt?: string,
   readonly score?: number,
   readonly ratings?: number,
@@ -25,35 +24,35 @@ interface ProfileProps {
 }
 
 const Profile = ({
-  className, medium, title, picture, checked, alt,
+  className, isMedium, title, picture, isIdChecked, alt,
   ratingsLabel, info, action, score = 0, ratings = 0,
 }: ProfileProps) => (
-  <div className={cc(['kirk-profile', prefix({ medium }), className])}>
-    <div className="kirk-description">
-      <span className="kirk-title">{title}</span>
-      { ratings > 0 &&
-        <Rating ratings={ratings} score={score}>{ratingsLabel}</Rating>
-      }
-      <span className="kirk-secondaryInfo">{info}</span>
-    </div>
-    { !isEmpty(picture) &&
-      <Avatar
-        image={picture}
-        alt={alt}
-        checked={checked}
-        className="kirk-picture"
-        medium={medium}
-      />
-    }
-    { !isEmpty(action) &&
-      <div>
-        <Button status={Button.STATUS.UNSTYLED} title={action}>
-          <ArrowIcon right />
-        </Button>
+    <div className={cc(['kirk-profile', prefix({ medium: isMedium }), className])}>
+      <div className="kirk-description">
+        <span className="kirk-title">{title}</span>
+        {ratings > 0 &&
+          <Rating ratings={ratings} score={score}>{ratingsLabel}</Rating>
+        }
+        <span className="kirk-secondaryInfo">{info}</span>
       </div>
-    }
-    <style jsx>{style}</style>
-  </div>
-)
+      {picture &&
+        <Avatar
+          image={picture}
+          alt={alt}
+          isIdChecked={isIdChecked}
+          className="kirk-picture"
+          isMedium={isMedium}
+        />
+      }
+      {action &&
+        <div>
+          <Button status={Button.STATUS.UNSTYLED} title={action}>
+            <ArrowIcon right />
+          </Button>
+        </div>
+      }
+      <style jsx>{style}</style>
+    </div>
+  )
 
 export default Profile