+++ src/profile/story.tsx
@@ -16,8 +16,8 @@ stories.add(
     // tslint:disable-next-line
     picture={text('picture', 'https://pbs.twimg.com/profile_images/749446875162505218/6r6-9wDn.jpg')}
     alt={text('alternative text', 'Avatar')}
-    medium={boolean('medium size', false)}
-    checked={boolean('checked', false)}
+    isMedium={boolean('medium size', false)}
+    isIdChecked={boolean('isIdChecked', false)}
     ratings={number('rating', 10)}
     score={number('score', 2.5)}
     ratingsLabel={text('ratings label', 'ratings')}
@@ -29,7 +29,7 @@ stories.add(
   'Profile with action',
   withInfo('')(() => (<Profile
     title={text('name', 'Pepe le Pew')}
-    medium={boolean('medium size', false)}
+    isMedium={boolean('medium size', false)}
     info={text('secondary info', '')}
     action="/"
   />)),